requirements.txt
matplotlib>=3.8
networkx>=3.2
numpy>=1.26
scikit-learn>=1.4
umap-learn>=0.5.5
